fix small bugs

+++ cloudflare-python.py
@@ -1,10 +1,10 @@
-#!/usr/bin/env python3 
+#!/usr/bin/env python3
 import sys
 import os
 import requests
 import json
 
-import module_locator 
+import module_locator
 '''for find absolute path dir. see https://coderoad.ru/2632199/Как-получить-путь-к-текущему-исполняемому-файлу-в-Python#2632297
 or https://stackoverflow.com/questions/122327/how-do-i-find-the-location-of-my-python-site-packages-directory
 '''
@@ -33,18 +33,18 @@ def read_conf():
     except IOError:
         print('''
         Credential file is empty, please paste Your credentials
-        and script will generete creds.json autimaticaly, 
+        and script will generete creds.json autimaticaly,
         or You can manual copy from exampl, see help on github
         ''')
         a = input('Please type "X-Auth-Email" (simple: youremail@gmail.com): ')
-        b = input('Please type "X-Auth-Key" (simple: 3210a271xxxxx413d2e8dexxxxxe1ffxxxxca): ') 
+        b = input('Please type "X-Auth-Key" (simple: 3210a271xxxxx413d2e8dexxxxxe1ffxxxxca): ')
         try:
             write_conf(a,b)
             sys.exit(0)
         except Exception as e:
             print('variables is lost, try another', e)
             sys.exit(1)
-    
+
     headers['X-Auth-Email'] = data['X-Auth-Email']
     headers['X-Auth-Key']   = data['X-Auth-Key']
 
@@ -53,11 +53,11 @@ read_conf()
 
 
 
-def add(zone,record,ipaddr): 
+def add(zone,record,ipaddr):
     data = {
         'type':     "A",
-        'ttl':      "120", 
-        'priority': 10, 
+        'ttl':      "120",
+        'priority': 10,
         'proxied':  False
     }
     data['name']    = record
@@ -74,17 +74,19 @@ def add(zone,record,ipaddr):
         print('Exceprion:\n', e)
         print(r.json())
 
+
 def get_zone_id(zone):
     url = 'https://api.cloudflare.com/client/v4/zones'
     r = requests.get(url, headers=headers,  json='')
-    for i in r.json()['result']:
+    ress = r.json()['result']
+    for i in ress:
         if i.get('name') == zone:
             return i.get('id')
 
-def get_a_dns_records(zone): 
+def get_a_dns_records(zone):
     url = 'https://api.cloudflare.com/client/v4/zones/{}/dns_records'.format(str(get_zone_id(zone)))
     r = requests.get(url, headers=headers, json='')
- 
+
     list = []
     for i in r.json().get('result'):
         type = i.get('type')
@@ -108,22 +110,27 @@ def get_a_dns_records(zone):
     return r.json()
 
 def get_dns_record_id(zone,record):
-    url = 'https://api.cloudflare.com/client/v4/zones/{}/dns_records'.format(str(get_zone_id(zone)))
+    zone_id = str(get_zone_id(zone))
+    url = f'https://api.cloudflare.com/client/v4/zones/{zone_id}/dns_records'
     r = requests.get(url, headers=headers, json='')
-    all_records = r.json() 
+    all_records = r.json()
     for i in all_records.get('result'):
         name = i['name']
         id   = i['id']
         if name == record:
             return id
 
-def delete(record,zone):
-    print('now delete {}'.format(record))
+def def_delete(record,zone):
+    print(f'now delete {record}:')
     zone_id = str(get_zone_id(zone))
     dns_record_id = get_dns_record_id(zone,record)
-    url = 'https://api.cloudflare.com/client/v4/zones/{}/dns_records/{}'.format(zone_id,dns_record_id)
-    r = requests.delete(url, headers=headers, json='')
-    print(r.json())
+    if dns_record_id is not None:
+        url = f'https://api.cloudflare.com/client/v4/zones/{zone_id}/dns_records/{dns_record_id}'
+        r = requests.delete(url, headers=headers, json='')
+        print(r.json())
+
+    else:
+       print(f'subdomain {record} not exist!')
 
 def get_all_zones_list():
     url = 'https://api.cloudflare.com/client/v4/zones'
@@ -134,8 +141,13 @@ def get_all_zones_list():
         print(r.json().get('errors')[0].get('error_chain'))
         sys.exit(1)
     else:
-        for i in r.json()['result']:
-            print(i['name'])
+        domains_list = []
+        for i in r.json().get('result'):
+            if i is not None:
+                domains_list.append(i.get('name'))
+        return domains_list
+
+
 
 def help():
     print('./cloudflare-python.py add       subdomain.yourDomain.com 8.8.8.8')
@@ -146,7 +158,8 @@ def help():
 
 if len(sys.argv) == 2:
    if sys.argv[1] == 'get_zones':
-       get_all_zones_list()
+       for i in get_all_zones_list():
+           print(f'{i}')
    else:
        help()
 
@@ -156,7 +169,7 @@ elif len(sys.argv) == 4:
     ipaddr = sys.argv[3]
     zone   = '.'.join(record.split('.')[-2:])
     if mode == 'add':
-        add(zone,record,ipaddr)   
+        add(zone,record,ipaddr)
     else:
         help()
 
@@ -165,10 +178,20 @@ elif len(sys.argv) == 3:
     record = sys.argv[2]
     zone   = '.'.join(record.split('.')[-2:])
     if mode == 'del':
-        delete(record,zone)
+        if zone in get_all_zones_list():
+            def_delete(record,zone)
+        else:
+            print(f'Erro: zone "{zone}" does not exist in Your zones!')
+            sys.exit(1)
     elif mode == 'get_all_a':
-        get_a_dns_records(zone)
-    else: 
+        #get_a_dns_records(zone)
+        if zone in get_all_zones_list():
+            get_a_dns_records(zone)
+        else:
+            print(f'Erro: zone "{zone}" does not exist in Your zones!')
+            sys.exit(1)
+    else:
         help()
 else:
     help()
+
